+++ packages/ub-pub/CHANGELOG.md
@@ -6,8 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
- - `UB.Reposiroty` fabric function can accept serialized UBQL as a parameter (instead of entity code).
- In this case result will be a Repository de-serialized rom a given UBQL
+ - `UB.Reposiroty` fabric function can now accept a UBQL passed in parameter as object while keeping an ability
+ to pass entity code as string
  
  ``` 
    // serialize Repository into plain java script object (UBQL)